+++ x/model/client/cli/tx_model_version.go
@@ -4,6 +4,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/tx"
+	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
 	"github.com/zigbee-alliance/distributed-compliance-ledger/utils/cli"
 	"github.com/zigbee-alliance/distributed-compliance-ledger/x/model/types"
@@ -185,6 +186,48 @@ func CmdUpdateModelVersion() *cobra.Command {
 	_ = cmd.MarkFlagRequired(FlagPid)
 	_ = cmd.MarkFlagRequired(FlagSoftwareVersion)
 	_ = cmd.MarkFlagRequired(flags.FlagFrom)
+
+	return cmd
+}
+
+func CmdDeleteModelVersion() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "delete-model-version [vid] [pid] [software-version]",
+		Short: "Broadcast message DeleteModelVersion",
+		Args:  cobra.ExactArgs(3),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argVid, err := cast.ToInt32E(args[0])
+			if err != nil {
+				return err
+			}
+			argPid, err := cast.ToInt32E(args[1])
+			if err != nil {
+				return err
+			}
+			argSoftwareVersion, err := cast.ToInt32E(args[2])
+			if err != nil {
+				return err
+			}
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgDeleteModelVersion(
+				clientCtx.GetFromAddress().String(),
+				argVid,
+				argPid,
+				argSoftwareVersion,
+			)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
 
 	return cmd
 }
